rework to use a subclassed SystemdService resource
