+++ lib/chef/resource/service.rb
@@ -24,18 +24,16 @@ class Chef
     class Service < Chef::Resource
       identity_attr :service_name
 
-      state_attrs :enabled, :running, :masked
+      state_attrs :enabled, :running
 
       default_action :nothing
-      allowed_actions :enable, :disable, :start, :stop, :restart, :reload,
-                      :mask, :unmask
+      allowed_actions :enable, :disable, :start, :stop, :restart, :reload
 
       def initialize(name, run_context = nil)
         super
         @service_name = name
         @enabled = nil
         @running = nil
-        @masked = nil
         @parameters = nil
         @pattern = service_name
         @start_command = nil
@@ -137,15 +135,6 @@ class Chef
       def running(arg = nil)
         set_or_return(
           :running,
-          arg,
-          :kind_of => [ TrueClass, FalseClass ]
-        )
-      end
-
-      # if the service is masked or not
-      def masked(arg=nil)
-        set_or_return(
-          :masked,
           arg,
           :kind_of => [ TrueClass, FalseClass ]
         )
